+++ cmd/brabble/main.go
@@ -35,6 +35,7 @@ func run() error {
 	root.AddCommand(control.NewTestHookCmd(cfgPath))
 	root.AddCommand(control.NewDoctorCmd(cfgPath))
 	root.AddCommand(control.NewServiceCmd(cfgPath))
+	root.AddCommand(control.NewSetupCmd(cfgPath))
 
 	// Hidden internal serve command used by start.
 	root.AddCommand(daemon.NewServeCmd(cfgPath))