+++ internal/run/util.go
@@ -1,0 +1,8 @@
+package run
+
+func max(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}